requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
bs4>=0.0.1
python-dotenv>=1.0.0
pytz>=2023.3
lxml>=4.9.3
pytest>=7.4.0
